\name{BPSC-package}
\alias{BPSC-package}
\alias{BPSC}
\docType{package}
\title{
Beta-Poisson Model for Single-Cell RNA-seq Data Analyses
}
\description{
Beta-Poisson model for single-cell RNA-seq data analyses
}
\details{
\tabular{ll}{
Package: \tab BPSC\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2018-08-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Trung Nghia Vu, Mattias Rantalainen and Yudi Pawitan

Maintainer: Trung Nghia Vu <nghiavtr@gmail.com>

}
\examples{ 
# See vignette
vignette("BPSC")
}

\keyword{ package }


