% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimateBP.R
\name{estimateBP}
\alias{estimateBP}
\title{Estimate parameters of a beta-Poisson model for a data vector}
\usage{
estimateBP(x, para.num = 4, tbreak.num = 10, break.thres = 10,
  useExt = FALSE, param0 = NULL)
}
\arguments{
\item{x}{A vector of input data points}

\item{para.num}{Mode of beta-Poisson model: 3, 4 (default) or 5 parameters}

\item{tbreak.num}{Number of breaks for binning}

\item{break.thres}{A parameter setting of \code{\link{getTbreak}} function}

\item{useExt}{A parameter setting of \code{\link{getTbreak}} function that allows to extend the last bin to infinity or not}

\item{param0}{Initial parameters for the model. If it is not set, default initial parameters will be obtained from \code{\link{getInitParam}} function}
}
\value{
An optimal model to the input data including optimal parameters (par), X2 test results (X2 and PVAL), etc..
}
\description{
Estimate parameters of a beta-Poisson model for a data vector
}
\examples{
set.seed(2015)
#Create a simulated gene expression by randomly generating 100 data points
#from a beta-poisson model
alp=0.6;bet=1.5;lam1=20;lam2=0.05
par0=c(alp,bet,lam1,lam2)
bp.vec=rBP(100,par0)
#Estimate parameters of the four-parameter beta-Poisson model from the data points
res=estimateBP(bp.vec,para.num=4)
#Print the goodness-of-fit of the model and the optimal parameters
res$X2
res$par
}

