% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBin.R
\name{getBin}
\alias{getBin}
\title{Get bins of data points}
\usage{
getBin(x, tbreak = NULL, tbreak.num = NULL, minVal = 0, maxVal = NULL)
}
\arguments{
\item{x}{A vector of input data points}

\item{tbreak}{A vector of break points}

\item{tbreak.num}{Number of breaks}

\item{minVal}{Under bound of the binning range}

\item{maxVal}{Upper bound of the binning range}
}
\value{
A vector of bin values
}
\description{
Get bins of data points
}
\examples{
set.seed(2015)
bp.vec=rBP(100,alp=0.6,bet=1.5,lam1=20,lam2=0.05)
# get 10 bins for the data
getBin(bp.vec,tbreak.num=10)
# get bins from a predefined tbreak
tbreak=getTbreak(bp.vec,tbreak.num=10,break.thres=10)
getBin(bp.vec,tbreak=tbreak)
}

