% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBPMCnull.R
\name{getBPMCnull}
\alias{getBPMCnull}
\title{Generate a Monte-Carlo null distribution for a beta-Poisson model}
\usage{
getBPMCnull(param, n = 200, tbreak = NULL, tbreak.num = 20,
  sim.num = 1000, ran.num = 1e+05, E.esp = 0, useDebug = FALSE)
}
\arguments{
\item{param}{A vector of parameters of the beta-Poisson model}

\item{n}{The number of data points for each simulation}

\item{tbreak}{A vector of break points}

\item{tbreak.num}{Number of breaks}

\item{sim.num}{A number of simulations}

\item{ran.num}{The number of data points generated from the beta-Poisson model to approximate the theoretical model}

\item{E.esp}{An small value added to expected value when computing X2}

\item{useDebug}{A parameter just used for debug and checking, so useDebug=FALSE by default}
}
\value{
A list containning the Monte-Carlo null distribution and other values
}
\description{
Generate a Monte-Carlo null distribution for a beta-Poisson model
}
\examples{
set.seed(2015)
par=c(0.6,1.5,20,0.05)
MCnull.res=getBPMCnull(par,n=100,sim.num=100)
hist(MCnull.res$X2,xlab="goodness-of-fit statistic",
main="Monte-Carlo null distribution",breaks=20, prob=TRUE)
}

