% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getInitParam.R
\name{getInitParam}
\alias{getInitParam}
\title{Estimate initial parameters of the beta-Poisson model for a data vector}
\usage{
getInitParam(x, para.num = 4)
}
\arguments{
\item{x}{A vector of input data points}

\item{para.num}{Mode of beta-Poisson model: 3, 4 (default) or 5 parameters}
}
\value{
Initial parameters for BP models of the data points
}
\description{
Estimate initial parameters of the beta-Poisson model for a data vector
}
\examples{
set.seed(2015)
bp.vec=rBP(100,alp=0.6,bet=1.5,lam1=20,lam2=0.05)
init.par=getInitParam(bp.vec,para.num=4)
}

