% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getMCpval.R
\name{getMCpval}
\alias{getMCpval}
\title{Get Monte-Carlo p-values from the BP models and the Monte-Carlo null distributions}
\usage{
getMCpval(bp.model.list = NULL, MCdis.list = NULL)
}
\arguments{
\item{bp.model.list}{List of beta-Poisson models that are results from \code{\link{estimateBPMatrix}} function}

\item{MCdis.list}{List of Monte-Carlo null distributions that are results of \code{\link{getBPMCnullmatrix}} function}
}
\value{
A list of Monte-Carlo p-values
}
\description{
Get Monte-Carlo p-values from the BP models and the Monte-Carlo null distributions
}
\examples{
#See \\code{\\link{getBPMCnullmatrix}} function
}

