% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTbreak.R
\name{getTbreak}
\alias{getTbreak}
\title{Get breaks for binning of data points}
\usage{
getTbreak(x, tbreak.num = 10, break.thres = 10, useExt = FALSE,
  useQuantile = FALSE)
}
\arguments{
\item{x}{A vector of input data points}

\item{tbreak.num}{Number of breaks}

\item{break.thres}{A threshold to decide whether use breaks from logs cale or not. We decide to use breaks from log scale if 75 percent of input data less than this threshold}

\item{useExt}{An option to extend the last bin to infinity or not}

\item{useQuantile}{An option to bin the data points by equal-size quantile values}
}
\value{
A vector of break values
}
\description{
Get breaks for binning of data points
}
\examples{
set.seed(2015)
bp.vec=rBP(100,alp=0.6,bet=1.5,lam1=20,lam2=0.05)
tbreak=getTbreak(bp.vec,tbreak.num=10)
}

