% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pBP.R
\name{pBP}
\alias{pBP}
\title{Compute the probability distribution function (pdf) of a single quantile from a four-parameter beta-Poisson model}
\usage{
pBP(x, alp, bet, lam1 = 1, lam2 = 1)
}
\arguments{
\item{x}{A non-negative quantile}

\item{alp}{A non-negative value, the parameter of Beta function (alpha)}

\item{bet}{A non-negative value, the parameter of Beta function (beta)}

\item{lam1}{A non-negative value, the parameter for scaling}

\item{lam2}{A non-negative value, the parameter for smoothing, used for BP4 or BP5}
}
\value{
pdf of the input value
}
\description{
Compute the probability distribution function (pdf) of a single quantile from a four-parameter beta-Poisson model
}
\examples{
x=0.3
pBP(x,alp=0.6,bet=1.5,lam1=20,lam2=0.05)
}

