% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pBPi.R
\name{pBPi}
\alias{pBPi}
\title{Compute the probability distribution function (pdf) of an interval from a four-parameter beta-Poisson model}
\usage{
pBPi(x1, x2, alp, bet, lam1 = 1, lam2 = 1)
}
\arguments{
\item{x1}{Under bound of the interval}

\item{x2}{Upper bound of the interval}

\item{alp}{A non-negative value, the parameter of Beta function (alpha)}

\item{bet}{A non-negative value, the parameter of Beta function (beta)}

\item{lam1}{A non-negative value, the parameter for scaling}

\item{lam2}{A non-negative value, the parameter for smoothing, used for BP4 or BP5}
}
\value{
pdf of the interval
}
\description{
Compute the probability distribution function (pdf) of an interval from a four-parameter beta-Poisson model
}
\examples{
x1=0.2
x2=0.4
pBPi(x1,x2,alp=0.6,bet=1.5,lam1=20,lam2=0.05)
}

