% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rBP.R
\name{rBP}
\alias{rBP}
\title{Generate random values from a beta-Poisson model}
\usage{
rBP(n, alp, bet, lam1 = 1, lam2 = 1, prob0 = 0)
}
\arguments{
\item{n}{The number of random data points}

\item{alp}{A non-negative value, the parameter of Beta function (alpha)}

\item{bet}{A non-negative value, the parameter of Beta function (beta)}

\item{lam1}{A non-negative value, the parameter for scaling}

\item{lam2}{A non-negative value, the parameter for smoothing, used for BP4 or BP5}

\item{prob0}{A non-negative value in [0;1), the parameter for the portion of fixed density of zero in the distribution, used for BP5}
}
\value{
A vector of the generated data points
}
\description{
Generate random values from a beta-Poisson model
}
\examples{
set.seed(2015)
#BP3
bp.vec=rBP(100,0.2,1.5,200)
hist(bp.vec, prob=TRUE)
#BP4
bp.vec=rBP(100,0.6,1.5,20,0.05)
hist(bp.vec, prob=TRUE)
#BP5
bp.vec=rBP(100,0.2,1.5,200,0.1,0.05)
hist(bp.vec, prob=TRUE)
}

