% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.BPglm.R
\name{summary.BPglm}
\alias{summary.BPglm}
\title{Summarize results of beta-Poisson generalized linear models of a dataset}
\usage{
\method{summary}{BPglm}(object, ...)
}
\arguments{
\item{object}{An object of BPglm class resulted from \code{\link{BPglm}} function}

\item{...}{Inherited parameters of summary object}
}
\value{
An summary.BPglm object contains a table of coefficients sorted by estimated false discovery rate (FDR)
}
\description{
Summarize results of beta-Poisson generalized linear models of a dataset
}
\examples{
#See the example in the \\code{\\link{BPglm}} function
}

